fix to fix

+++ lang/cem/libcc.ansi/stdlib/atoi.c
@@ -10,7 +10,7 @@
    overflow.
 */
 int
-atol(register const char *nptr)
+atoi(register const char *nptr)
 {
 	int total = 0;
 	register unsigned int digit;
